Finally proper handling of Power Dialog override settings (1 of 2)
Summary:
The easiest way is to introduce another BrodacastReceier. When override
is enabled, we take down the alert-producing receiver and register our
receiver. Our receiver simply records values but does not act on them.
That way, if alerts are enabled again, the values are up to date.

Test Plan: works, but test when we're done with upstream merges

Reviewers: synergy, anders3408, jriordan001, raymanfx, donald.munn, rlasalle15

Reviewed By: synergy

Differential Revision: http://review.codefirex.com/D469