Emulate stock behavior of key events that don't originate from hard keys
This is getting a bit tricky now, but I think this should do it.
We already don't touch any back or menu button virtual key events
For Home and AppSwitcher, if a navbar is visible, we simulate a single
tap by executing the correct behavior through the ActionHandler and
consume the event.

Change-Id: Id62559b123ea5ad96b21a4b506b8fe1d49b7d657