bootanimation: Allow devices to avoid loading many textures
The bootanimation attempts to guess whether or not it is safe
to load all the textures for a "part" into memory to speed
up boot time.

This is causing problems on some devices and therefore
allow devices to select the slower (but safer) mode
where it will not load all the textures into memory.

Change-Id: If18686261d024c410cc355c1f90001b9449fd0bf