NxPulse: Build NxPulse with android-visualizer library new abstraction
This flushes out all the reduntant classes that are part of the library
and allows NxPulse operations to be confined to a single class. In addition,
we took the opportunity to tighten up NxModule and its callbacks. Ultimately,
the idea is to create a robust "module" interface. However, i'm not certain
how many NxModules will come to fruition.

Change-Id: I01cb26c2c39fd0dc4f155140d3648f1045909765