Pulse: add isVisualizerLocked API call to AudioManager
When Fling and Pulse are enabled, this creates a lock on
Visualizer instances. In this condition, only system packages
can create Visualizer instances. When a app is locked out, safely
throw UnsupportedOperationException since caller must catch it
anyways. This protects Pulse from losing visualizer link from
user apps that use visualizer api. This is needed since Pulse
is in a persistent state. We can internally manage other
instances of visualizer in system.

Change-Id: I508c1d5474eaa78d90d8337adcc17b4c70ae7166